utils/requirements.txt
pydantic==1.10.8
Pympler==1.0.1
PyYAML==6.0.1